requirements.txt
rich
llama-cpp-python
